uagents_core/requirements.txt
uagents>=0.22.0
uagents-core>=0.4.4
httpx>=0.27.0
supabase>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
